chore: Rename CurrentTenant provider to CurrentTenantProvider
