--- packages/webapp-libs/webapp-tenants/src/providers/currentTenant/index.ts
@@ -1,4 +1,0 @@
-export * from './currentTenant.types';
-export * from './currentTenant.component';
-export * from './currentTenant.graphql';
-export * from './currentTenant.hook';
