+++ packages/webapp-libs/webapp-tenants/src/providers/currentTenantProvider/currentTenantProvider.graphql.ts
