+++ packages/webapp-libs/webapp-tenants/src/providers/currentTenantProvider/index.ts
@@ -1,0 +1,4 @@
+export * from './currentTenantProvider.types';
+export * from './currentTenantProvider.component';
+export * from './currentTenantProvider.graphql';
+export * from './currentTenantProvider.hook';
