+++ packages/webapp-libs/webapp-tenants/src/providers/index.ts
@@ -1,1 +1,1 @@
-export * from './currentTenant';
+export * from './currentTenantProvider';
